requirements.txt
asgiref==3.8.1
Django~=5.0.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
django-monthfield==0.1.4
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1
django-cron==0.6.0
django-countries==7.6.1
djangorestframework-simplejwt==5.3.1
requests~=2.32.3
celery~=5.4.0
django-celery-beat==2.6.0
psycopg2==2.9.9
django-ipware==7.0.1
XlsxWriter~=3.2.0
urllib3~=2.2.2
openpyxl~=3.1.5
django-filter~=24.3
django-celery-results==2.5.1
celery==5.4.0